migrate tests for result check
